Formatting/Typo tweaks
