+++ HacktoberfestProject.Web/Startup.cs
@@ -1,17 +1,12 @@
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
-
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-
-using HacktoberfestProject.Web.Extensions.DependancyInjection;
 using HacktoberfestProject.Web.Data.Configuration;
 using HacktoberfestProject.Web.Data;
-using System;
-using HacktoberfestProject.Web.Testing;
-using System.Runtime.CompilerServices;
+using HacktoberfestProject.Web.Extensions.DependencyInjection;
 
 namespace HacktoberfestProject.Web
 {
